Ambient Occlusion / smooth Lighting mostly works now!

+++ client/src/voxel/meshgen/shared.h
@@ -9,6 +9,6 @@ uint chunkGetGeneratedThisFrame();
 
 sideMask chunkGetSides              (u16 x,u16 y,u16 z,blockId b[CHUNK_SIZE+2][CHUNK_SIZE+2][CHUNK_SIZE+2]);
 void     chunkPopulateBlockData     (blockId b[CHUNK_SIZE+2][CHUNK_SIZE+2][CHUNK_SIZE+2], chunk *c, i16 xoff, i16 yoff, i16 zoff);
-void     chunkOptimizePlane         (u64 plane[CHUNK_SIZE][CHUNK_SIZE]);
+void     chunkOptimizePlane         (u32 plane[CHUNK_SIZE][CHUNK_SIZE]);
 uint     chunkGetGeneratedThisFrame ();
 void     chunkResetCounter          ();
